requirements.txt
google-generativeai
streamlit
langchain
langchain-google-genai
chromadb
pypdf
tiktoken
